cs65_4/cs65_5/requirements.txt
django
django-cors-headers
django-apscheduler
djangorestframework
pymongo
bs4
furl
requests
